linker: don't fail if text relocations are found
Yes, the correct solution would have been to fix the relocations
in the programs affected by them, but sometimes this isn't
possible, e.g. if we're given a precompiled shared library which
has text relocations and we cannot rebuild it.

So, as a last hack to get things going, just skip the validation
in the linker and go on with just the plain old warning.

Signed-off-by: Aleksander Morgado <aleksander@aleksander.es>